Don't do a full get_transfer every 5 seconds
There is code here to skip it if the balance hasn't changed, but it is
suspiciously commented out (and according to git blame has been that way
for a very long time).  On large wallets this is causing massive CPU
usage, wallet lag, and may be responsible for the balance update and
no-staked-nodes issues.